Scrolling works! :D
Index.tsx: imported useWindowDimensions and const height and width based on screen size, so that this parent element now has a dynamic size it's children can reference to.

PackingList.tsx: the scroll in FlatList was stopped by the style on the internal Pressable. To override this, I added contentContainerStyle to the FlatList element. I also imported height from useWindowDimensions and added a padding so that text will not be too long before new line.